Beggining our scheduler
This is the first commit to our scheduler using sched_ext framework.

This scheduler is been developed by PATOS to a Hackathon of Isovalent in
2025. Our idea is to build a simplified scheduler to prioritize specific
containers in kubernetes.

The callbacks `init`, `exit`, `enable` are done. Some other have to take
a look.

For now, our idea is to configre a simpolified CFS with more  nice value
in the containers processes that we want to prioritize