Add redis manifest deploy
Since Redis container was used on benchmark tests, it makes sense to add
in the repository our manifests to the public.

Note that the container is deployed with a visible port on the host
(30001), because we did not want to use a Ingress/Egress load balancing
(e.g. Traefik) to lead with visibility of the container. This probably
would cause more unnecessary configuration and one more process to
compete with Redis and Stress-noise.